Every time when fuel_data is updated - insert anchors
This is required in situations when we are fetching partially
changed graph from fuel, and there is already commited state in
solar.
For example, if we will add role to already commited node -
changes that are specific for a role will be disjoined, and may
run in arbitrary order.

However this change doesnt guarantee total correctness of order,
if there will be no events that point to/from stage, and newly added
resources will rely on commited resource to establish order - we will
end up with disjoined graph.